feat: refactor map character placement logic and add geojson utilities
